Minor error fixes

+++ R/clear_f1_cache.R
@@ -27,7 +27,9 @@ clear_f1_cache <- function() {
     }
   }
 
-  unlink(file.path(getOption("f1dataR.cache"), "f1dataR_http_cache"), recursive = TRUE)
+  if (getOption("f1dataR.cache") %in% c("off", "memory")) {
+    unlink(file.path(getOption("f1dataR.cache"), "f1dataR_http_cache"), recursive = TRUE)
+  }
 
   memoise::forget(f1dataR::load_drivers)
   memoise::forget(f1dataR::load_laps)
